+++ azure_monitor/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+## 0.6b.0
+Released 2021-01-28
+
+- Remove request metrics from auto-collection
+  ([#124](https://github.com/microsoft/opentelemetry-azure-monitor-python/pull/124))
+- Implement standard metrics for http dependency telemetry
+  ([#125](https://github.com/microsoft/opentelemetry-azure-monitor-python/pull/125))
+
 ## 0.5b.0
 Released 2020-09-24